realized that i should have the functions in wordbox instead of app removed unused css
